+++ modules/demux/asf/asf.c
@@ -2,14 +2,14 @@
  * asf.c : ASFv01 file input module for vlc
  *****************************************************************************
  * Copyright (C) 2001 VideoLAN
- * $Id: asf.c,v 1.11 2002/12/03 17:00:16 fenrir Exp $
+ * $Id: asf.c,v 1.12 2002/12/06 16:34:06 sam Exp $
  * Authors: Laurent Aimar <fenrir@via.ecp.fr>
- * 
+ *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation; either version 2 of the License, or
  * (at your option) any later version.
- * 
+ *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
@@ -56,7 +56,7 @@ static uint16_t GetWLE( uint8_t *p_buff )
     return( (p_buff[0]) + ( p_buff[1] <<8 ) );
 }
 static uint32_t GetDWLE( uint8_t *p_buff )
-{   
+{
     return( p_buff[0] + ( p_buff[1] <<8 ) +
             ( p_buff[2] <<16 ) + ( p_buff[3] <<24 ) );
 }
@@ -65,14 +65,14 @@ static uint32_t GetDWLE( uint8_t *p_buff )
  * Activate: check file and initializes ASF structures
  *****************************************************************************/
 static int Activate( vlc_object_t * p_this )
-{   
+{
     input_thread_t  *p_input = (input_thread_t *)p_this;
     uint8_t         *p_peek;
     guid_t          guid;
-    
+
     demux_sys_t     *p_demux;
     int             i_stream;
-    
+
     /* Initialize access plug-in structures. */
     if( p_input->i_mtu == 0 )
     {
@@ -96,7 +96,7 @@ static int Activate( vlc_object_t * p_this )
     }
 
     /* create our structure that will contains all data */
-    if( !( p_input->p_demux_data = 
+    if( !( p_input->p_demux_data =
                 p_demux = malloc( sizeof( demux_sys_t ) ) ) )
     {
         msg_Err( p_input, "out of memory" );
@@ -120,7 +120,7 @@ static int Activate( vlc_object_t * p_this )
         msg_Warn( p_input, "ASF v1.0 plugin discarded (not a valid file)" );
         return( -1 );
     }
-   
+
     if( !( p_demux->p_fp = ASF_FindObject( p_demux->root.p_hdr,
                                     &asf_object_file_properties_guid, 0 ) ) )
     {
@@ -129,7 +129,7 @@ static int Activate( vlc_object_t * p_this )
         msg_Warn( p_input, "ASF v1.0 plugin discarded (missing file_properties object)" );
         return( -1 );
     }
-    
+
     if( p_demux->p_fp->i_min_data_packet_size != p_demux->p_fp->i_max_data_packet_size )
     {
         ASF_FreeObjectRoot( p_input, &p_demux->root );
@@ -137,8 +137,8 @@ static int Activate( vlc_object_t * p_this )
         msg_Warn( p_input, "ASF v1.0 plugin discarded (invalid file_properties object)" );
         return( -1 );
     }
-    
-    p_demux->i_streams = ASF_CountObject( p_demux->root.p_hdr, 
+
+    p_demux->i_streams = ASF_CountObject( p_demux->root.p_hdr,
                                           &asf_object_stream_properties_guid );
     if( !p_demux->i_streams )
     {
@@ -174,19 +174,19 @@ static int Activate( vlc_object_t * p_this )
     {
         asf_stream_t    *p_stream;
         asf_object_stream_properties_t *p_sp;
-        
+
         p_sp = ASF_FindObject( p_demux->root.p_hdr,
                                &asf_object_stream_properties_guid,
                                i_stream );
 
-        p_stream = 
-            p_demux->stream[p_sp->i_stream_number] = 
+        p_stream =
+            p_demux->stream[p_sp->i_stream_number] =
                 malloc( sizeof( asf_stream_t ) );
         memset( p_stream, 0, sizeof( asf_stream_t ) );
         p_stream->p_sp = p_sp;
 
         vlc_mutex_lock( &p_input->stream.stream_lock );
-        p_stream->p_es = 
+        p_stream->p_es =
             input_AddES( p_input,
                          p_input->stream.p_selected_program,
                          p_sp->i_stream_number,
@@ -207,42 +207,42 @@ static int Activate( vlc_object_t * p_this )
 
             p_stream->i_cat = AUDIO_ES;
             msg_Dbg( p_input,
-                    "adding new audio stream(codec:0x%x,ID:%d)", 
+                    "adding new audio stream(codec:0x%x,ID:%d)",
                     i_codec,
                     p_sp->i_stream_number );
             switch( i_codec )
             {
                 case( 0x01 ):
-                    p_stream->p_es->i_fourcc = 
+                    p_stream->p_es->i_fourcc =
                         VLC_FOURCC( 'a', 'r', 'a', 'w' );
                     break;
                 case( 0x50 ):
                 case( 0x55 ):
-                    p_stream->p_es->i_fourcc = 
+                    p_stream->p_es->i_fourcc =
                         VLC_FOURCC( 'm','p','g','a' );
                     break;
                 case( 0x2000 ):
-                    p_stream->p_es->i_fourcc = 
+                    p_stream->p_es->i_fourcc =
                         VLC_FOURCC( 'a','5','2',' ' );
                     break;
                 case( 0x160 ):
-                    p_stream->p_es->i_fourcc = 
+                    p_stream->p_es->i_fourcc =
                         VLC_FOURCC( 'w','m','a','1' );
                     break;
                 case( 0x161 ):
-                    p_stream->p_es->i_fourcc = 
+                    p_stream->p_es->i_fourcc =
                         VLC_FOURCC( 'w','m','a','2' );
                     break;
                 default:
-                    p_stream->p_es->i_fourcc = 
+                    p_stream->p_es->i_fourcc =
                         VLC_FOURCC( 'm','s',(i_codec >> 8)&0xff,i_codec&0xff );
             }
             if( p_sp->i_type_specific_data_length > 0 )
             {
                 WAVEFORMATEX    *p_wf;
-                int             i_size;
+                size_t          i_size;
                 uint8_t         *p_data;
-                
+
                 i_size = p_sp->i_type_specific_data_length;
 
                 p_wf = malloc( i_size );
@@ -264,14 +264,14 @@ static int Activate( vlc_object_t * p_this )
                             i_size - sizeof( WAVEFORMATEX ) );
                 }
             }
-            
+
         }
         else
         if( CmpGUID( &p_sp->i_stream_type, &asf_object_stream_type_video ) )
         {
             p_stream->i_cat = VIDEO_ES;
             msg_Dbg( p_input,
-                    "adding new video stream(ID:%d)", 
+                    "adding new video stream(ID:%d)",
                     p_sp->i_stream_number );
             if( p_sp->p_type_specific_data )
             {
@@ -283,21 +283,21 @@ static int Activate( vlc_object_t * p_this )
             }
             else
             {
-                p_stream->p_es->i_fourcc = 
+                p_stream->p_es->i_fourcc =
                     VLC_FOURCC( 'u','n','d','f' );
             }
             if( p_sp->i_type_specific_data_length > 11 )
             {
                 BITMAPINFOHEADER *p_bih;
-                int         i_size;
+                size_t      i_size;
                 uint8_t     *p_data;
 
                 i_size = p_sp->i_type_specific_data_length - 11;
-                
+
                 p_bih = malloc( i_size );
                 p_stream->p_es->p_demux_data = (void*)p_bih;
                 p_data = p_sp->p_type_specific_data + 11;
-                
+
                 p_bih->biSize       = GetDWLE( p_data );
                 p_bih->biWidth      = GetDWLE( p_data + 4 );
                 p_bih->biHeight     = GetDWLE( p_data + 8 );
@@ -322,8 +322,8 @@ static int Activate( vlc_object_t * p_this )
         else
         {
             p_stream->i_cat = UNKNOWN_ES;
-            msg_Dbg( p_input, 
-                    "ignoring unknown stream(ID:%d)", 
+            msg_Dbg( p_input,
+                    "ignoring unknown stream(ID:%d)",
                     p_sp->i_stream_number );
             p_stream->p_es->i_fourcc = VLC_FOURCC( 'u','n','d','f' );
         }
@@ -336,7 +336,7 @@ static int Activate( vlc_object_t * p_this )
         }
         vlc_mutex_unlock( &p_input->stream.stream_lock );
     }
-    
+
 
     p_demux->i_data_begin = p_demux->root.p_data->i_object_pos + 50;
     if( p_demux->root.p_data->i_object_size != 0 )
@@ -361,12 +361,12 @@ static int Activate( vlc_object_t * p_this )
         mtime_t i_length;
 
         /* real number of packets */
-        i_count = ( p_input->stream.p_selected_area->i_size - 
-                       p_demux->i_data_begin ) / 
+        i_count = ( p_input->stream.p_selected_area->i_size -
+                       p_demux->i_data_begin ) /
                             p_demux->p_fp->i_min_data_packet_size;
         /* calculate the time duration in s */
-        i_length = (mtime_t)p_demux->p_fp->i_play_duration / 10 * 
-                   (mtime_t)i_count / 
+        i_length = (mtime_t)p_demux->p_fp->i_play_duration / 10 *
+                   (mtime_t)i_count /
                    (mtime_t)p_demux->p_fp->i_data_packets_count /
                    (mtime_t)1000000;
         if( i_length > 0 )
@@ -391,12 +391,12 @@ static int Activate( vlc_object_t * p_this )
     p_input->stream.p_selected_program->b_is_ok = 1;
     vlc_mutex_unlock( &p_input->stream.stream_lock );
 
-    
+
     return( 0 );
 }
 
 /*****************************************************************************
- * Demux: read packet and send them to decoders 
+ * Demux: read packet and send them to decoders
  *****************************************************************************/
 #define GETVALUE2b( bits, var, def ) \
     switch( (bits)&0x03 ) \
@@ -419,13 +419,13 @@ static int Demux( input_thread_t *p_input )
         off_t i_offset;
         msleep( DEFAULT_PTS_DELAY );
         i_offset = ASF_TellAbsolute( p_input ) - p_demux->i_data_begin;
-        
+
         if( i_offset  < 0 )
         {
             i_offset = 0;
         }
         /* XXX work only when i_min_data_packet_size == i_max_data_packet_size */
-        i_offset += p_demux->p_fp->i_min_data_packet_size - 
+        i_offset += p_demux->p_fp->i_min_data_packet_size -
                         i_offset % p_demux->p_fp->i_min_data_packet_size;
         ASF_SeekAbsolute( p_input, p_demux->i_data_begin + i_offset );
 
@@ -448,7 +448,7 @@ static int Demux( input_thread_t *p_input )
         int     i_data_packet_min = p_demux->p_fp->i_min_data_packet_size;
         uint8_t *p_peek;
         int     i_skip;
-        
+
         int     i_packet_size_left;
         int     i_packet_flags;
         int     i_packet_property;
@@ -457,14 +457,14 @@ static int Demux( input_thread_t *p_input )
         int     i_packet_length;
         int     i_packet_sequence;
         int     i_packet_padding_length;
-        
+
         uint32_t    i_packet_send_time;
         uint16_t    i_packet_duration;
         int         i_payload;
         int         i_payload_count;
         int         i_payload_length_type;
 
-        
+
         if( input_Peek( p_input, &p_peek, i_data_packet_min ) < i_data_packet_min )
         {
             // EOF ?
@@ -472,7 +472,7 @@ static int Demux( input_thread_t *p_input )
             return( 0 );
         }
         i_skip = 0;
-        
+
         /* *** parse error correction if present *** */
         if( p_peek[0]&0x80 )
         {
@@ -484,7 +484,7 @@ static int Demux( input_thread_t *p_input )
             i_opaque_data_present = ( p_peek[0] >> 4 )& 0x01;    // 1bit
             i_error_correction_length_type = ( p_peek[0] >> 5 ) & 0x03; // 2bits
             i_skip += 1; // skip error correction flags
-        
+
             if( i_error_correction_length_type != 0x00 ||
                 i_opaque_data_present != 0 ||
                 i_error_correction_data_length != 0x02 )
@@ -501,7 +501,7 @@ static int Demux( input_thread_t *p_input )
 
         i_packet_flags = p_peek[i_skip]; i_skip++;
         i_packet_property = p_peek[i_skip]; i_skip++;
-        
+
         b_packet_multiple_payload = i_packet_flags&0x01;
 
         /* read some value */
@@ -511,11 +511,11 @@ static int Demux( input_thread_t *p_input )
 
         i_packet_send_time = GetDWLE( p_peek + i_skip ); i_skip += 4;
         i_packet_duration  = GetWLE( p_peek + i_skip ); i_skip += 2;
-    
+
 //        i_packet_size_left = i_packet_length;   // XXX données reellement lu
         /* FIXME I have to do that for some file, I don't known why */
         i_packet_size_left = i_data_packet_min;
-        
+
         if( b_packet_multiple_payload )
         {
             i_payload_count = p_peek[i_skip] & 0x3f;
@@ -544,13 +544,13 @@ static int Demux( input_thread_t *p_input )
             mtime_t i_pts;
             mtime_t i_pts_delta;
 
-            i_stream_number = p_peek[i_skip] & 0x7f; 
+            i_stream_number = p_peek[i_skip] & 0x7f;
             i_skip++;
-            
+
             GETVALUE2b( i_packet_property >> 4, i_media_object_number, 0 );
             GETVALUE2b( i_packet_property >> 2, i_tmp, 0 );
             GETVALUE2b( i_packet_property, i_replicated_data_length, 0 );
-            
+
             if( i_replicated_data_length > 1 ) // should be at least 8 bytes
             {
                 i_pts = (mtime_t)GetDWLE( p_peek + i_skip + 4 ) * 1000;
@@ -584,25 +584,25 @@ static int Demux( input_thread_t *p_input )
             }
             else
             {
-                i_payload_data_length = i_packet_length - 
+                i_payload_data_length = i_packet_length -
                                             i_packet_padding_length - i_skip;
             }
-            
+
 #if 0
-             msg_Dbg( p_input, 
+             msg_Dbg( p_input,
                       "payload(%d/%d) stream_number:%d media_object_number:%d media_object_offset:%d replicated_data_length:%d payload_data_length %d",
-                      i_payload + 1, 
+                      i_payload + 1,
                       i_payload_count,
                       i_stream_number,
-                      i_media_object_number, 
-                      i_media_object_offset, 
-                      i_replicated_data_length, 
+                      i_media_object_number,
+                      i_media_object_offset,
+                      i_replicated_data_length,
                       i_payload_data_length );
 #endif
 
             if( !( p_stream = p_demux->stream[i_stream_number] ) )
             {
-                msg_Warn( p_input, 
+                msg_Warn( p_input,
                           "undeclared stream[Id 0x%x]", i_stream_number );
                 i_skip += i_payload_data_length;
                 continue;   // over payload
@@ -615,9 +615,9 @@ static int Demux( input_thread_t *p_input )
             }
 
 
-            for( i_payload_data_pos = 0; 
-                 i_payload_data_pos < i_payload_data_length && 
-                        i_packet_size_left > 0; 
+            for( i_payload_data_pos = 0;
+                 i_payload_data_pos < i_payload_data_length &&
+                        i_packet_size_left > 0;
                  i_payload_data_pos += i_sub_payload_data_length )
             {
                 data_packet_t  *p_data;
@@ -645,7 +645,7 @@ static int Demux( input_thread_t *p_input )
 
                 if( !p_stream->p_pes )  // add a new PES
                 {
-                    p_stream->i_time = 
+                    p_stream->i_time =
                         ( (mtime_t)i_pts + i_payload * (mtime_t)i_pts_delta );
 
                     if( p_demux->i_first_pts == -1 )
@@ -655,9 +655,9 @@ static int Demux( input_thread_t *p_input )
                     p_stream->i_time -= p_demux->i_first_pts;
 
                     p_stream->p_pes = input_NewPES( p_input->p_method_data );
-                    p_stream->p_pes->i_dts = 
-                        p_stream->p_pes->i_pts = 
-                            input_ClockGetTS( p_input, 
+                    p_stream->p_pes->i_dts =
+                        p_stream->p_pes->i_pts =
+                            input_ClockGetTS( p_input,
                                               p_input->stream.p_selected_program,
                                               ( p_stream->i_time+DEFAULT_PTS_DELAY) * 9 /100 );
 
@@ -675,7 +675,7 @@ static int Demux( input_thread_t *p_input )
                 p_data->p_payload_start += i_skip;
                 i_packet_size_left -= i_read;
 
-                
+
                 if( !p_stream->p_pes->p_first )
                 {
                     p_stream->p_pes->p_first = p_stream->p_pes->p_last = p_data;
@@ -700,7 +700,7 @@ static int Demux( input_thread_t *p_input )
                 }
             }
         }
-        
+
         if( i_packet_size_left > 0 )
         {
             if( !ASF_SkipBytes( p_input, i_packet_size_left ) )
@@ -744,14 +744,14 @@ loop_error_recovery:
     {
         /* update pcr XXX in mpeg scale so in 90000 unit/s */
         p_demux->i_pcr =( __MAX( p_demux->i_time /*- DEFAULT_PTS_DELAY*/, 0 ) ) * 9 / 100;
-        
+
         /* first wait for the good time to read next packets */
         input_ClockManageRef( p_input,
                               p_input->stream.p_selected_program,
                               p_demux->i_pcr );
     }
 
-   
+
     return( 1 );
 }
 
@@ -759,14 +759,14 @@ loop_error_recovery:
  * MP4End: frees unused data
  *****************************************************************************/
 static void Deactivate( vlc_object_t * p_this )
-{   
+{
 #define FREE( p ) \
-    if( p ) { free( p ); } 
-   
+    if( p ) { free( p ); }
+
     input_thread_t *  p_input = (input_thread_t *)p_this;
     demux_sys_t *p_demux = p_input->p_demux_data;
     int i_stream;
-    
+
     msg_Dbg( p_input, "Freeing all memory" );
     ASF_FreeObjectRoot( p_input, &p_demux->root );
     for( i_stream = 0; i_stream < 128; i_stream++ )
@@ -782,7 +782,7 @@ static void Deactivate( vlc_object_t * p_this )
         }
 #undef p_stream
     }
-    
+
 #undef FREE
 }
